tests: move utilities under a test_framework module
