+++ tests/fixtures.py
@@ -5,7 +5,7 @@ of the file i originally copied! (MIT licensed)
 """
 from concurrent import futures
 from ephemeral_port_reserve import reserve
-from utils import (
+from test_framework.utils import (
     BitcoinD,
     ManagerRevaultd,
     StakeholderRevaultd,
